Added @extend feature. Closes #149
Credit for the idea goes to SASS. For example you may
now:

form button
  padding: 10px
  border-radius: 5px
  border: 1px solid gray

a.button
  @extends form button
  text-decoration: none

form
  button
    padding: 10px
    color: red
    &:hover
      color: blue

yielding:

form button,
a.button {
  padding: 10px;
  border-radius: 5px;
  border: 1px solid #808080;
}
a.button {
  text-decoration: none;
}
form button {
  padding: 10px;
  color: #f00;
}
form button:hover {
  color: #00f;
}

This is a bit more powerful than the SASS implementation (AFAIK)
as you are not required to use single-segment selectors.

Known issues:

  - docs needed
  - when extending with no properties present it will not work (yet)
  - missing "multiple" inheritance

you may use @extend or @extends, they are the same.

One thing we could potentially add/improve upon is
perhaps altering the lookup but its most likely fine
as it is here