Fix middleware overlap function on windows.
The `compare` function was comparing file system paths, and url paths, but figuring out which separator to use based on the OS path separator `sep`.  This did not work on windows, and no overlap would be found.  

This fixes the compare function, and the part of the middleware that uses the overlap output so that the compare function translates the os path that might overlap with the http path into the correct path using a separator that can be compared to an url path.

This fixes windows compatibility, and unix/linux still works fine.

I will submit a test in a moment.