Added support for UTF-8 encoding
I was working on a mixin to inline SVG's with the ability to set `fill` colors. Inlining of SVG's performs better when we skip the base64 encoding, which wasn't yet enabled in Stylus, like it is in LESS for example.

With my update you can just call `embedurl("path/to/file.svg", "utf8")` (or whatever you've defined as the alias for Stylus' url method) and it'll return the `url(` string you can use with `background-image`.

- Added a test & updated the docs
- Fixed some issues that were causing tests to fail on Windows.
- All test files with `url` were using Stylus' url method when it didn't need to so I changed the `define` from `url` to `embedurl`.