Use less confusing variable names
Whilst the original shows how you can compose variables and mix this with other keywords to make a property declaration:
```
body
  font font, san-serif
```
This will be confusing to anyone unfamiliar with Stylus. It would also not be generally considered good practice to name a variable the same as the property it's applied to.

Theses possibilities and technicalities are better discovered later by more advanced users, than in the documentation on variables which is likely to be seen and used by every beginner.